requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.3.4
google-auth==2.34.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.143.0
requests==2.32.3
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart==0.0.9
jinja2==3.1.4
itsdangerous==2.2.0